Appathon/requirements.txt
langchain-core>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
pandas>=2.0.0
python-dotenv>=1.0.0
openai>=1.12.0